# docs/api.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Hired. backend",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/setup/test-provider": {
      "post": {
        "tags": [
          "setup"
        ],
        "summary": "Test Provider Endpoint",
        "description": "Run a tiny round-trip against the requested provider.",
        "operationId": "test_provider_endpoint_api_setup_test_provider_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TestProviderRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestProviderResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/setup/select-provider": {
      "post": {
        "tags": [
          "setup"
        ],
        "summary": "Select Provider",
        "description": "Commit the user's provider choice. Persists app_config + keychain.",
        "operationId": "select_provider_api_setup_select_provider_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SelectProviderRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SelectProviderResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SelectProviderRequest": {
        "properties": {
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "api_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Key"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          }
        },
        "type": "object",
        "required": [
          "provider"
        ],
        "title": "SelectProviderRequest"
      },
      "SelectProviderResponse": {
        "properties": {
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          }
        },
        "type": "object",
        "required": [
          "provider"
        ],
        "title": "SelectProviderResponse"
      },
...
      "TestProviderRequest": {
        "properties": {
          "provider": {
            "type": "string",
            "title": "Provider",
            "description": "One of: mock, anthropic_api, claude_code, ollama"
          },
          "api_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Key"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          }
        },
        "type": "object",
        "required": [
          "provider"
        ],
        "title": "TestProviderRequest"
      },
      "TestProviderResult": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "latency_ms": {
            "type": "integer",
            "title": "Latency Ms"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "error_kind": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "missing_api_key",
                  "auth_failed",
                  "rate_limited",
                  "network_error",
                  "bad_response",
                  "model_unavailable",
                  "binary_missing",
                  "unknown",
                  "unsupported_provider"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Kind"
          }
        },
        "type": "object",
        "required": [
          "ok",
          "latency_ms",
          "error",
          "error_kind"
        ],
        "title": "TestProviderResult"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}